Updated the auto-command, to escape special chars
Apparently, one needs to distinguish between Windows and Unix, because
on Windows you have to replace all backslashes by forward slashed or
otherwise the auto-command won't trigger. This obviously creates
problems on unix, when editing files, that contain a backslash in their
path, which could easily be the case. Therefore, in the function that
sets up the SwapBRP auto-command, we are now distinguishing these 2
cases.

Additionally, the balloonexpr has been built into the autocommand. This
allows vim to show a little Popup that shows, whether the mouse is over
the modified/recovered version or the on-disk version of the file.

And a unused variable has been deleted.