Improve getting vim's recovery warning message
With Neovim the old way of parsing gives terminal codes in the output. Vim can save its verbose output to a file, so use that instead.
I think a nice side effect is that this approach also supports Windows, but I don't have a Windows machine to test it on.