system('uname') might be slow
Shelling out to `uname` and testing whether the operating system is
linux is slow, especially since this is done everytime when vim starts
up. So instead rely on the existence of the /proc directory and assume
it is linux in that case. The existence of that directory was actually
the only reason to test for a linux system in the first place

closes #67