handle BSD stat correctly
Or better, do not try to use BSD stat, which doesn't know the `--printf`
argument (which is only available for GNU stat)

Dear reader: If you have a bsd stat available, can you please send a
patch that makes use of bsd stat -f format? Or just send me a way to
just output the owner and the last modification time. Thank you!

closes #73